fix typo

+++ include/physics_system.hpp
@@ -9,7 +9,7 @@
 class PhysicsSystem {
    public:
     void step(float elapsed_ms);
-    static bool collidesPolyVec(Entity island_entity, vec2 node_pos);
+    static bool collidesPolyVec(Entity island_entity, ivec2 node_pos);
 
     PhysicsSystem() {}
 };
